// src/FE/src/api/generated/models/inventoryMaterialResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Workslip.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface InventoryMaterialResponse {
  id: string;
  name: string;
  sku: string;
  unit: string;
  /** @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$ */
  unitCost: number | string;
  isActive: boolean;
  qrCode: string;
}